[chore] Separated Financial data provider files
Cleaned up gulpfile
Added task for component folders
