+++ package.json
@@ -9,6 +9,7 @@
     "gulp": "^3.5.6",
     "gulp-bump": "~0.1.8",
     "gulp-concat": "^2.2.0",
+    "gulp-folders": "^1.0.0",
     "gulp-html2js": "^0.1.0",
     "gulp-jshint": "~1.5.5",
     "gulp-minify-css": "^0.3.5",
