+++ src/financial/real-time.js
@@ -1,53 +1,8 @@
 var RiseVision = RiseVision || {};
 RiseVision.Common = RiseVision.Common || {};
+RiseVision.Common.Financial = RiseVision.Common.Financial || {};
 
-RiseVision.Common.Financial = {};
-RiseVision.Common.Financial.Helper = {};
 RiseVision.Common.Financial.RealTime = {};
-RiseVision.Common.Financial.Historical = {};
-RiseVision.Common.Financial.Historical.CollectionTimes = {};
-
-RiseVision.Common.Financial.Helper = function(instruments) {
-  this.instruments = instruments;
-};
-
-RiseVision.Common.Financial.Helper.prototype.setInstruments = function(instruments) {
-  this.instruments = instruments;
-};
-
-RiseVision.Common.Financial.Helper.prototype.getInstruments = function(isLoading, collectionTimes) {
-  var self = this;
-
-  if (isLoading) {
-    return this.instruments.join("|");
-  }
-  else {
-    var dayOfWeek = new Date().getDay(), len = collectionTimes.length, instruments = [];
-
-    $.each(this.instruments, function(i, instrument) {
-      for (var j = 0; j < len; j++) {
-        if (instrument === collectionTimes[j].instrument) {
-          var startTime = collectionTimes[j].startTime, endTime = collectionTimes[j].endTime, daysOfWeek = collectionTimes[j].daysOfWeek;
-
-          //Check if the instrument should be requested again based on its collection data.
-          $.each(daysOfWeek, function(j, day) {
-            //Check collection day.
-            if (day === dayOfWeek) {
-              //Check collection time.
-              if (new Date().between(startTime, endTime)) {
-                instruments.push(self.instruments[i]);
-              }
-
-              return false;
-            }
-          });
-        }
-      }
-    });
-
-    return instruments.join("|");
-  }
-};
 
 RiseVision.Common.Financial.RealTime = function(displayID, instruments) {
   var self = this;
@@ -421,253 +376,3 @@ RiseVision.Common.Financial.RealTime.prototype.saveBeforeValue = function(field,
     });
   }
 };
-
-RiseVision.Common.Financial.Historical = function(displayID, instrument, duration) {
-  var self = this;
-
-  if (displayID) {
-    this.displayID = displayID;
-  }
-  else {
-    this.displayID = "preview";
-  }
-
-  this.instrument = instrument;
-  this.duration = duration;
-  this.isLoading = true;
-  this.updateInterval = 60000;
-  this.now = Date.today();
-  //Issue 922
-  this.url = "http://contentfinancial2.appspot.com/data/historical?";
-  this.historicalViz = new RiseVision.Common.Visualization();
-  this.helper = new RiseVision.Common.Financial.Helper([this.instrument]);
-};
-
-RiseVision.Common.Financial.Historical.prototype.setInstrument = function(instrument) {
-  this.isLoading = true;
-  this.instrument = instrument;
-  this.helper.setInstruments([this.instrument]);
-};
-
-RiseVision.Common.Financial.Historical.prototype.setDuration = function(duration) {
-  this.duration = duration;
-};
-
-RiseVision.Common.Financial.Historical.prototype.setIsUpdated = function(isUpdated) {
-  CollectionTimes.getInstance().setIsUpdated(this.instrument, isUpdated);
-};
-/* Historical Financial data - Only one stock can be requested at a time. */
-RiseVision.Common.Financial.Historical.prototype.getHistoricalData = function(fields, callback, options) {
-  var self = this, queryString = "select " + fields.join() + " ORDER BY tradeTime", codes = "";
-
-  //Customize the query string.
-  if (options) {
-    if (options.sortOrder) {
-      if (options.sortOrder === "desc") {
-        queryString += " desc";
-      }
-    }
-
-    if (options.limit) {
-      queryString += " LIMIT " + options.limit;
-    }
-  }
-
-  CollectionTimes.getInstance().addInstrument(this.instrument, this.now, function(times, now) {
-    self.now = now;
-    codes = self.helper.getInstruments(self.isLoading, [times]);
-
-    //Perform a search for the instrument.
-    if (codes) {
-      options = {
-        url : self.url + "id=" + self.displayID + "&code=" + self.instrument + "&kind=" + self.duration,
-        refreshInterval : 0,
-        queryString : queryString,
-        callback : function histCallback(data) {
-          self.onHistoricalDataLoaded(data, times, callback);
-        }
-      };
-
-      //Start a timer in case there is a problem loading the data (i.e. Internet has been disconnected).
-      self.getHistoricalDataTimer = setTimeout(function() {
-        self.getHistoricalData(fields, callback, options);
-      }, self.updateInterval);
-
-      self.historicalViz.getData(options);
-    }
-    //Request has been made outside of collection times.
-    else {
-      callback(null);
-    }
-  });
-};
-
-RiseVision.Common.Financial.Historical.prototype.onHistoricalDataLoaded = function(data, times, callback) {
-  var numDataRows = 0;
-
-  if (data !== null) {
-    clearTimeout(this.getHistoricalDataTimer);
-
-    this.historicalData = data;
-    numDataRows = data.getNumberOfRows();
-
-    if ((numDataRows === 0) || ((numDataRows === 1) && (data.getFormattedValue(0, 0) === "0"))) {
-      this.isLoading = true;
-    }
-    else {
-      this.isLoading = false;
-    }
-
-    if (this.historicalData !== null) {
-      callback({
-        "data" : this.historicalData,
-        "collectionData" : times
-      });
-    }
-    else {
-      callback({
-        "collectionData" : times
-      });
-    }
-  }
-  //Timeout or some other error occurred.
-  else {
-    console.log("Error encountered loading historical data for: ");
-    console.log(this);
-  }
-};
-
-/*
- * Singleton object to handle retrieving collection times for a historical instrument.
- */
-var CollectionTimes = (function() {
-  //Private variables and functions.
-  var instantiated = false, instruments = [];
-
-  function init() {
-    //Issue 903 Start
-    function loadCollectionTimes(instrument, callback) {
-      var updateInterval = 60000, viz = new RiseVision.Common.Visualization(), options;
-
-      //Start a timer in case there is a problem loading the data (i.e. Internet has been disconnected).
-      collectionTimesTimer = setTimeout(function() {
-        loadCollectionTimes(instrument, callback);
-      }, updateInterval);
-
-      options = {
-        //Change me for Production.
-        url : "http://contentfinancial2.appspot.com/info?codes=" + instrument,
-        refreshInterval : 0,
-        queryString : "select startTime, endTime, daysOfWeek, timeZoneOffset, updateInterval",
-        callback : function(result, timer) {
-          viz = null;
-
-          if (result !== null) {
-            clearTimeout(timer);
-            saveCollectionTimes(instrument, result);
-            callback();
-          }
-          //Timeout or some other error occurred.
-          else {
-            console.log("Error encountered loading collection times for: " + instrument);
-          }
-        },
-        params : collectionTimesTimer
-      };
-
-      viz.getData(options);
-    }
-
-    //Issue 903 End
-
-    function saveCollectionTimes(instrument, data) {
-      var numRows, startTime, endTime, timeZoneOffset;
-
-      if (data !== null) {
-        numRows = data.getNumberOfRows();
-
-        for (var i = 0; i < instruments.length; i++) {
-          if (instruments[i].instrument === instrument) {
-            timeZoneOffset = data.getValue(0, 3);
-            startTime = data.getValue(0, 0);
-            endTime = data.getValue(0, 1);
-
-            instruments[i].collectionTimes = {
-              "instrument" : instrument,
-              "startTime" : startTime.setTimezoneOffset(timeZoneOffset),
-              "endTime" : endTime.setTimezoneOffset(timeZoneOffset),
-              "daysOfWeek" : data.getFormattedValue(0, 2).split(","),
-              "timeZoneOffset" : timeZoneOffset,
-              "isUpdated" : true
-            };
-
-            break;
-          }
-        }
-      }
-    }
-
-    return {
-      setIsUpdated : function(instrument, isUpdated) {
-        for (var i = 0; i < instruments.length; i++) {
-          if (instruments[i].instrument === instrument) {
-            if (instruments[i].collectionTimes !== null) {
-              instruments[i].collectionTimes.isUpdated = isUpdated;
-            }
-          }
-        }
-      },
-      addInstrument : function(instrument, now, callback) {
-        var i = 0, instrumentFound = false, collectionTimesFound = false;
-
-        //Check if there is already collection data for this instrument.
-        for (; i < instruments.length; i++) {
-          if (instruments[i].instrument === instrument) {
-            //Issue 922 Start
-            if (instruments[i].collectionTimes !== null) {
-              if ((!Date.equals(Date.today(), now)) && (!instruments[i].collectionTimes.isUpdated)) {
-                now = Date.today();
-                instruments[i].collectionTimes.startTime.addDays(1);
-                instruments[i].collectionTimes.endTime.addDays(1);
-                instruments[i].collectionTimes.isUpdated = true;
-              }
-
-              collectionTimesFound = true;
-            }
-            //Issue 922 End
-
-            instrumentFound = true;
-            break;
-          }
-        }
-
-        if (collectionTimesFound) {
-          callback(instruments[i].collectionTimes, now);
-        }
-        else {
-          if (!instrumentFound) {
-            instruments.push({
-              instrument : instrument,
-              collectionTimes : null
-            });
-          }
-
-          loadCollectionTimes(instrument, function() {
-            callback(instruments[i].collectionTimes, now);
-          });
-        }
-      }
-    };
-  }
-
-  //Public functions.
-  return {
-    getInstance : function() {
-      if (!instantiated) {
-        instantiated = init();
-      }
-
-      return instantiated;
-    }
-  };
-})();
